refactor(DonutChart): reorganize type imports and enhance type definitions for clarity
